+++ graphix/sim/density_matrix.py
@@ -19,7 +19,7 @@ from graphix import linalg_validations as lv
 from graphix import parameter
 from graphix.channels import KrausChannel
 from graphix.parameter import Expression, ExpressionOrFloat, ExpressionOrSupportsComplex
-from graphix.sim.base_backend import DenseState, FullStateBackend, Matrix, kron, matmul, outer, tensordot, vdot
+from graphix.sim.base_backend import DenseState, DenseStateBackend, Matrix, kron, matmul, outer, tensordot, vdot
 from graphix.sim.statevec import CNOT_TENSOR, CZ_TENSOR, SWAP_TENSOR, Statevec
 from graphix.states import BasicStates, State
 
@@ -400,7 +400,7 @@ class DensityMatrix(DenseState):
 
 
 @dataclass(frozen=True)
-class DensityMatrixBackend(FullStateBackend[DensityMatrix]):
+class DensityMatrixBackend(DenseStateBackend[DensityMatrix]):
     """MBQC simulator with density matrix method."""
 
     state: DensityMatrix = dataclasses.field(init=False, default_factory=lambda: DensityMatrix(nqubit=0))